third_party/blink/web_tests/virtual/scoped-custom-element-registry-disabled/external/wpt/custom-elements/registries/Element-innerHTML-expected.txt

This is a testharness.js-based test.
[FAIL] innerHTML on a disconnected element should use the scoped registry it was created with
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] nested descendants in innerHTML on a disconnected element should use the scoped registry the element was created with
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] innerHTML on a disconnected element should use the scoped registry it was created with when parsing a simple HTML
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] innerHTML on an inserted element should continue to use the scoped registry it was created with
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] nested descendants in innerHTML should use the null registry when the container element has null registry
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] insertAdjacentHTML should use the element's registry even when the registry is null
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] createContextualFragment on a range inside a template should use null registry even when the template has a scoped registry
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] createContextualFragment on a range inside an element with scoped registry should use the scoped registry of the element
  Failed to construct 'CustomElementRegistry': Illegal constructor
Harness: the test ran to completion.
